feat: get users from server
