+++ src/app/users/data-access/types/user.interface.ts
@@ -1,28 +1,3 @@
-export interface UserInterface {
-	id: number
-	name: string
-	username: string
-	email: string
-	address: AddressInterface
-	phone: string
-	website: string
-	company: CompanyInterface
-}
-interface AddressInterface {
-	street: string
-	suite: string
-	city: string
-	zipcode: string
-	geo: GeoInterface
-}
+import { UserResponseInterface } from './user-response.interface'
 
-interface GeoInterface {
-	lat: string
-	lng: string
-}
-
-interface CompanyInterface {
-	name: string
-	catchPhrase: string
-	bs: string
-}
+export interface UserInterface extends UserResponseInterface {}
