+++ src/app/users/data-access/users-api.service.ts
@@ -1,16 +1,22 @@
-import { Observable } from 'rxjs'
+import { catchError, EMPTY, Observable } from 'rxjs'
 import { Injectable } from '@angular/core'
 import { HttpClient } from '@angular/common/http'
 
-import { UserInterface } from './types/user.interface'
+import { UserResponseInterface } from './types/user-response.interface'
 import { environment } from '../../../environments/environment.development'
 
 @Injectable()
 export class UsersApiService {
 	constructor(private http: HttpClient) {}
 
-	getUsers(): Observable<UserInterface[]> {
+	fetchUsers(): Observable<UserResponseInterface[]> {
 		const url = `${environment.apiUrl}/users`
-		return this.http.get<UserInterface[]>(url)
+
+		return this.http.get<UserResponseInterface[]>(url).pipe(
+			catchError((error) => {
+				console.error(error)
+				return EMPTY
+			}),
+		)
 	}
 }
