+++ src/app/users/data-access/users.service.ts
@@ -1,5 +1,15 @@
-import { UserInterface } from './types/user.interface'
+import { Injectable } from '@angular/core'
 
+import { UserInterface } from './types/user.interface'
+import { UsersApiService } from './users-api.service'
+
+@Injectable()
 export class UsersService {
 	users: UserInterface[] = []
+
+	constructor(private usersApiService: UsersApiService) {}
+
+	getUsers() {
+		this.usersApiService.fetchUsers().subscribe((users: UserInterface[]) => (this.users = users))
+	}
 }
