+++ src/app/users/feature/users-list/users-list.component.ts
@@ -1,10 +1,19 @@
-import { Component } from '@angular/core'
+import { Component, OnInit } from '@angular/core'
+
+import { UsersService } from '../../data-access/users.service'
 
 @Component({
 	selector: 'app-users-list',
 	standalone: true,
 	imports: [],
+	providers: [UsersService],
 	templateUrl: './users-list.component.html',
 	styleUrl: './users-list.component.scss',
 })
-export class UsersListComponent {}
+export class UsersListComponent implements OnInit {
+	constructor(private usersService: UsersService) {}
+
+	ngOnInit(): void {
+		this.usersService.getUsers()
+	}
+}
